+++ tests/testthat/test-04-get_files_content-GitLab.R
@@ -280,3 +280,12 @@ test_that("get_files_content makes use of files_structure", {
     with_cols = "api_url"
   )
 })
+
+test_that("get_files_content skips if files structure is empty", {
+  expect_snapshot(
+    files_content <- gitlab_testhost_priv$get_files_content_from_files_structure(
+      files_structure = test_mocker$use("gl_empty_files_structure"),
+      verbose = TRUE
+    )
+  )
+})